ai-engine/requirements.txt
web3==6.15.1
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
openai==1.12.0
anthropic==0.18.1
schedule==1.2.1
redis==5.0.1
pymongo==4.6.1
websockets==12.0
aiohttp==3.9.3
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
